Separate general utils methods from discord utils methods
