+++ main.py
@@ -5,7 +5,7 @@ import discord
 from discord.ext import commands
 from dotenv import load_dotenv
 from datetime import datetime
-from src import utils
+from src import utils, discord_utils
 
 
 def main():
@@ -48,11 +48,11 @@ async def run(bot: commands.Bot, logger: logging.Logger):
     """
     logger.info("Setting up private server object")
     try:
-        utils.SERVER = discord.Object(id=int(os.getenv("PRIVATE_SERVER_SUPPORT_ID")))
+        discord_utils.SERVER = discord.Object(id=int(os.getenv("PRIVATE_SERVER_SUPPORT_ID")))
         logger.info("Private server set!")
     except TypeError:
         logger.info("Could not set private server object, please be wary that owner commands are usable everywhere")
-        utils.SERVER = None
+        discord_utils.SERVER = None
 
     for filename in os.listdir('src/Scripts'):
         if filename.endswith('py'):
